Agent action space & initialization fix
Changed so that AI can only move as many steps in positive or negative x & y directions as the amount of tiles on the board. E.g., if there are 5 tiles on the board, the agent starts at the center tile and can move 5 steps in any direction when trying to place tiles. Fixed observation for this and observation space. Also reworked initialization of AI.